v.0.0.6 - Separate server, ui and app .R files
So far I had the functions for the ui and the server in the same file. It is a bit annoying to work like this, separating them helps finding what you need. I do not know if posit connect cloud supports having them separated, so this is kind of a test.